-Added webpage reading support to both command parsers, to allow for auto updating bots/flythroughs
